move dish connection status to the sidebar
