+++ srt/dashboard/layouts/system_page.py
@@ -4,8 +4,6 @@ Function for Generating System Page and Creating Callback
 
 """
 
-import dash
-
 try:
     from dash import dcc
 except:
@@ -31,43 +29,6 @@ def generate_layout(config=None):
     -------
     System Page Layout
     """
-    is_bigdish = config is not None and config.get("MOTOR_TYPE") == "W1XMBIGDISH"
-
-    bigdish_panel = html.Div(
-        [
-            html.Div(
-                [
-                    html.H4(
-                        "Antenna Session",
-                        style={"text-align": "center"},
-                    ),
-                    dcc.Markdown(id="bigdish-status"),
-                    html.Div(
-                        [
-                            dbc.Button(
-                                "Connect",
-                                id="bigdish-connect-btn",
-                                color="primary",
-                                className="me-2",
-                                style={"margin-right": "8px"},
-                            ),
-                            dbc.Button(
-                                "Kick & Connect",
-                                id="bigdish-kick-btn",
-                                color="danger",
-                            ),
-                        ],
-                        style={"text-align": "center", "margin-top": "8px"},
-                    ),
-                    html.Div(id="bigdish-connect-result"),
-                ],
-                className="pretty_container four columns",
-            ),
-        ],
-        className="flex-display",
-        style={"justify-content": "center", "margin": "5px"},
-    ) if is_bigdish else html.Div()
-
     layout = html.Div(
         [
             html.Div(
@@ -159,13 +120,12 @@ def generate_layout(config=None):
                 className="flex-display",
                 style={"justify-content": "center", "margin": "5px"},
             ),
-            bigdish_panel,
         ]
     )
     return layout
 
 
-def register_callbacks(app, config, status_thread, command_thread=None):
+def register_callbacks(app, config, status_thread):
     """Registers the Callbacks for the System Page
 
     Parameters
@@ -183,8 +143,6 @@ def register_callbacks(app, config, status_thread, command_thread=None):
     -------
     None
     """
-    is_bigdish = config.get("MOTOR_TYPE") == "W1XMBIGDISH"
-
     @app.callback(
         Output("emergency-contact-info", "children"),
         [Input("interval-component", "n_intervals")],
@@ -275,36 +233,3 @@ def register_callbacks(app, config, status_thread, command_thread=None):
             else:
                 return [html.Li(html.A(filename)) for filename in (files + folders)]
 
-    if is_bigdish:
-        @app.callback(
-            Output("bigdish-status", "children"),
-            [Input("interval-component", "n_intervals")],
-        )
-        def update_bigdish_status(n):
-            status = status_thread.get_status()
-            if status is None:
-                return "Daemon not connected"
-            busy = status.get("bigdish_session_busy", False)
-            if busy:
-                return "**Session busy** — another client is active"
-            return "**Session active**"
-
-        @app.callback(
-            Output("bigdish-connect-result", "children"),
-            [
-                Input("bigdish-connect-btn", "n_clicks"),
-                Input("bigdish-kick-btn", "n_clicks"),
-            ],
-        )
-        def handle_bigdish_buttons(connect_clicks, kick_clicks):
-            ctx = dash.callback_context
-            if not ctx.triggered:
-                return ""
-            if command_thread is None:
-                return "No command thread available"
-            button_id = ctx.triggered[0]["prop_id"].split(".")[0]
-            if button_id == "bigdish-connect-btn":
-                command_thread.add_to_queue("bigdish_connect")
-            elif button_id == "bigdish-kick-btn":
-                command_thread.add_to_queue("bigdish_connect kick")
-            return ""
